dryrun in datate script

+++ bin/php/dataset.php
@@ -15,6 +15,7 @@ $options = $script->getOptions(
     '',
     array(
         'generate_from_class' => 'Genera un oggetto dataset in eZ dato p\'identificativo di classe specificato',
+        'dry-run' => 'Verifica l\azione ma non la esegue',
         'id' => 'eZ Content Object id',
         'show' => 'Mostra le info sul dataset',
         'push' => 'Salva o aggiorna il dataset in Ckan',
@@ -38,9 +39,15 @@ try {
     if ($options['generate_from_class']) {
         $generator = $tools->getDatasetGenerator();
         if ($generator instanceof OcOpendataDatasetGeneratorInterface) {
-            $object = $generator->createFromClassIdentifier($options['generate_from_class']);
+            $object = $generator->createFromClassIdentifier($options['generate_from_class'],
+                $options['dry-run'] !== null);
+            if (!$options['dry-run']) {
+                $cli->warning("Generato/aggiornato oggetto " . $object->attribute('id'));
+            } else {
+                $cli->warning('Ok');
+            }
         } else {
-            throw new Exception( 'Generator not found');
+            throw new Exception('Generator not found');
         }
     } else {
 
@@ -72,18 +79,30 @@ try {
         }
 
         if ($options['push']) {
-            $tools->pushObject($object);
-            $cli->warning('Push OK');
+            if (!$options['dry-run']) {
+                $tools->pushObject($object);
+                $cli->warning('Push OK');
+            } else {
+                $cli->warning('Ok');
+            }
         }
 
         if ($options['delete']) {
-            $tools->deleteObject($object);
-            $cli->warning('Delete OK');
+            if (!$options['dry-run']) {
+                $tools->deleteObject($object);
+                $cli->warning('Delete OK');
+            } else {
+                $cli->warning('Ok');
+            }
         }
 
         if ($options['purge']) {
-            $tools->deleteObject($object, true);
-            $cli->warning('Purge OK');
+            if (!$options['dry-run']) {
+                $tools->deleteObject($object, true);
+                $cli->warning('Purge OK');
+            } else {
+                $cli->warning('Ok');
+            }
         }
     }
